refactor: extract the associations layer into a mixin
The four methods under the "Associations" banner -- associated_logical_paths,
associated_pdsfiles, associated_abspaths and associated_parallel -- move verbatim
into src/pdsfile/_associations.py as _AssociationsMixin, which the class
statement takes as a base in its alphabetical slot, at the front. The bodies are
byte-for-byte what they were; nothing in this commit is a content edit inside a
moved body.

The window ends at associated_parallel, not at the end of the banner block. That
is the one boundary in this PR that does not follow a banner: is_logical_path is
a generic path predicate rather than an association, and it is a frozen public
PdsFile classmethod, so making it a plain module function would remove it from
the class surface. The parent plan says leave it in core and PR-22's stay-list
names it again, so it and the module-level tail below the class stay put.

The association rules do not move either: the ASSOCIATIONS translator and the
CATEGORIES set stay on PdsFile, where each rule subclass supplies its own.

The free-variable sweep names three module-level references: os, _clean_join and
_needs_glob. No class object appears, so this module needs no function-local
deferred import. _needs_glob loses its last caller in pdsfile.py, so its import
moves into the redundant-alias re-export block that already carries
_GLOB_CACHE_SIZE and _clean_glob -- deleting it would drop a name reachable today
as pdsfile.pdsfile._needs_glob, and an inline noqa and a ratchet grow are both
forbidden. os and _clean_join keep live callers and keep their plain form.

The new module's per-file-ignores entry is a split of pdsfile.py's: UP024, the
one code its moved lines carry, already forgiven for the same lines in the same
file, and every code in pdsfile.py's entry conserves.
